Minor changes + pubspec update
main_page.dart
Changed a const that wasn't supposed to be a const,
this prevented it from running on my pc.

for the register screen:
Removed and changed the "forgot password text"?
Changed the register button (formerly a login)
Removed the onclick route since it still led tot the initial login page.